refactor(api): extract resolveMergedTeamIds helper (audit consistency)
clutch and insights each inlined a byte-identical ~18-line block that resolves
a team's canonical ID and gathers every deprecated merged ID into a Set +
array. Extract to lib/team-merge.ts resolveMergedTeamIds(). Block moved
verbatim — behavior identical.

(The audit named 4 routes; in current code match-prediction/team-card don't do
merge resolution and the watchlist routes use team_merge_map for a different
purpose, so clutch + insights is the actual duplication.)

Verified: tsc clean, eslint clean, 295 tests pass.
