+++ frontend/app/api/insights/[teamId]/route.ts
@@ -1,4 +1,5 @@
 import { requirePremium } from '@/lib/api/requirePremium';
+import { resolveMergedTeamIds } from '@/lib/team-merge';
 import { isValidUuid } from '@/lib/validation';
 import { NextResponse } from 'next/server';
 import { generateAllInsights, type InsightInputData, type TeamInsightsResponse } from '@/lib/insights';
@@ -49,25 +50,8 @@ export async function GET(req: Request, { params }: { params: Promise<{ teamId: 
     }
 
     // Resolve merged team IDs so games stored under deprecated IDs are included
-    // (mirrors api.getTeamGames). Without this, streak/consistency stop at the
-    // first gap caused by games scraped under a pre-merge team_id.
-    const { data: incomingMerge } = await supabase
-      .from('team_merge_map')
-      .select('canonical_team_id')
-      .eq('deprecated_team_id', teamId)
-      .maybeSingle();
-    const canonicalTeamId = (incomingMerge as { canonical_team_id?: string } | null)?.canonical_team_id ?? teamId;
-
-    const { data: mergedTeams } = await supabase
-      .from('team_merge_map')
-      .select('deprecated_team_id')
-      .eq('canonical_team_id', canonicalTeamId);
-
-    const teamIdsToQuery = new Set<string>([canonicalTeamId, teamId]);
-    ((mergedTeams || []) as { deprecated_team_id: string | null }[]).forEach((m) => {
-      if (m.deprecated_team_id) teamIdsToQuery.add(m.deprecated_team_id);
-    });
-    const teamIdList = Array.from(teamIdsToQuery);
+    // (without this, streak/consistency stop at the first pre-merge gap).
+    const { teamIdsToQuery, teamIdList } = await resolveMergedTeamIds(supabase, teamId);
 
     // Fetch games with opponent rankings (across canonical + all merged team IDs)
     const orConditions = teamIdList
